LoRAEdit/requirements.txt
deepspeed==0.17.0
toml
transformers
diffusers>=0.32.1
datasets
pillow
sentencepiece
protobuf
peft
torch-optimi
tensorboard
tqdm
safetensors
bitsandbytes
imageio[ffmpeg]
av
einops
accelerate
loguru
flash-attn
omegaconf
iopath
termcolor
hydra-core
easydict
ftfy
pytorch-optimizer
wandb
gradio
imageio
fastapi==0.112.2
diffsynth
git+https://github.com/facebookresearch/sam2.git